+++ R/utils.R
@@ -39,14 +39,29 @@ names2 <- function(x) {
 
 #' Infix attribute accessor
 #'
+#' @description
+#'
+#' \Sexpr[results=rd, stage=render]{purrr:::lifecycle("soft-deprecated")}
+#'
+#' Please use the `%@%` operator exported in rlang. It has an
+#' interface more consistent with `@`: uses NSE, supports S4 fields,
+#' and has an assignment variant.
+#'
 #' @param x Object
 #' @param name Attribute name
 #' @export
 #' @name get-attr
+#' @keywords internal
 #' @examples
 #' factor(1:3) %@% "levels"
 #' mtcars %@% "class"
-`%@%` <- function(x, name) attr(x, name, exact = TRUE)
+`%@%` <- function(x, name) {
+  signal_soft_deprecated(paste_line(
+    "`%@%` is soft-deprecated as of purrr 0.3.0.",
+    "Please use the operator provided in rlang instead."
+  ))
+  attr(x, name, exact = TRUE)
+}
 
 
 #' Generate random sample from a Bernoulli distribution